refactor: extracted social media links into its own component to be reused for mobile and large screens
