+++ components/right-sidebar/SidebarSocialMedia.tsx
@@ -4,7 +4,7 @@ import { createPortal } from "react-dom";
 import Button from "../shared/ui/Button";
 import SocialMediaModal from "./SocialMediaModal";
 import { User, Social } from "@prisma/client";
-import { socialMediaIconList } from "@/lib/constants/socialMediaList";
+import SocialMediaLinks from "./SocialMediaLinks";
 
 const SidebarSocialMedia = ({
   user,
@@ -23,43 +23,11 @@ const SidebarSocialMedia = ({
       >
         Update social link
       </Button>
-      <h3 className="paragraph-2-bold border-b border-white-500 py-4">
+      <h3 className="paragraph-2-bold border-white-500 border-b py-4">
         Social Media Links
       </h3>
 
-      <div className="mr-1 flex flex-col">
-        {user.socialMedia ? (
-          user.socialMedia.map((social) => {
-            const filtered = socialMediaIconList.filter(
-              (icon) => social.type === icon.type
-            );
-            const { icon: Icon } = filtered[0];
-
-            return (
-              social.username && (
-                <a
-                  key={social.id}
-                  target="_blank"
-                  href={social.link ?? "No link found"}
-                  className="group flex flex-wrap items-center justify-start "
-                >
-                  <Icon
-                    size={30}
-                    className="group-hover:fill-white-300 group-hover:duration-300"
-                  />
-                  <p className="paragraph-2-regular text-white-300 group-hover:text-white-100  group-hover:duration-300">
-                    @{social.username}
-                  </p>
-                </a>
-              )
-            );
-          })
-        ) : (
-          <p className="paragraph-2-regular text-white-300">
-            Oops! No socials!
-          </p>
-        )}
-      </div>
+      <SocialMediaLinks socialMedia={user.socialMedia} />
 
       {open &&
         createPortal(
